+++ src/controllers/interactions.ts
@@ -14,6 +14,7 @@ import { respondWithMessageInEmbed, Status } from '../services/discord/respondTo
 import { HTTPError } from '../services/http';
 import { makeRemoveRoleFromGroupMenu, handleRemoveRoleFromGroupMenu } from '../services/discord/components/remove-role-from-group-menu';
 import { makeZoneMenu, makeOperandMenu, makeConditionMenu, handleCreateFFlogsRole } from '../services/discord/components/create-fflogs-role';
+import { handleCreateCustomGroup, makeCreateGroupMenu1, makeCreateGroupMenu2 } from '../services/discord/components/create-group-menu';
 
 export async function handleInteractions(req: Request, res: Response) {
     try {
@@ -74,6 +75,10 @@ export async function handleInteractions(req: Request, res: Response) {
                 case makeConditionMenu().custom_id:
                     [req, res] = await populateGuild(req, res);
                     return await handleCreateFFlogsRole(req, res);
+                case makeCreateGroupMenu1().custom_id:
+                case makeCreateGroupMenu2().custom_id:
+                    [req, res] = await populateGuild(req, res);
+                    return await handleCreateCustomGroup(req, res);
                 default:
                     return res.send({
                         type: InteractionResponseType.UPDATE_MESSAGE,